refactor: make utils/services dir
