+++ packages/portal/src/components/embed/EmbedGateway.vue
@@ -78,7 +78,7 @@
 
 <script>
   import klaroMixin from '@/mixins/klaro.js';
-  import serviceForUrl from '@/utils/getServiceForUrl';
+  import serviceForUrl from '@/utils/services/index.js';
 
   export default {
     name: 'EmbedGateway',
